backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Firebase
firebase-admin==6.6.0

# AI Services
groq==0.13.0
assemblyai==0.49.0
murf==1.1.1
google-search-results==2.4.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4

# Document Processing
python-docx==1.1.2

# Web & HTTP
requests==2.32.3
aiohttp==3.11.11
httpx==0.28.1

# WebSockets
websockets==14.1

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Report Generation
reportlab==4.2.5
jinja2==3.1.5
weasyprint==63.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Logging
loguru==0.7.3